+++ users/src/sdk/sdk.ts
@@ -53,9 +53,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0.0";
-    sdkVersion = "1.2.0";
-    genVersion = "2.169.0";
-    userAgent = "speakeasy-sdk/typescript 1.2.0 2.169.0 1.0.0 @wingspan/users";
+    sdkVersion = "1.3.0";
+    genVersion = "2.173.0";
+    userAgent = "speakeasy-sdk/typescript 1.3.0 2.173.0 1.0.0 @wingspan/users";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);